Finished forms management module for zushar v0.2.0 removed some utility functions after reconsideration of the feature list regarding form management and form-responses management fixed bugs in forms model and redesigned forms schema fixed typos in commenting of forms index file and also added more documentation to the forms module combined forms management utility functions with the CRUD forms model zushar api is now declared finished
